Optimizing Collision detection and pushing it into C++

+++ include/clock.hpp
@@ -8,14 +8,14 @@
 class Clock
 {
 private:
-	uint32_t lastframetime;
-	std::list<uint32_t> pastframesqueue;
+	float lastframetime;
+	std::list<float> pastframesqueue;
 
 public:
 	Clock(uint32_t starttime = 0);
 
 	uint32_t getTime();
-	uint32_t tick(uint32_t framerate = 0);
+	float tick(uint32_t framerate = 0);
 	float getFPS();
 };
